fastapi-rag-react/requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
openai
pydantic
pypdf
langchain
PyPDF2